% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{DADA2Adapter}
\alias{DADA2Adapter}
\title{Connect the DADA2 Pipeline to PreLect Process}
\usage{
DADA2Adapter(seqtab, taxa, out_path)
}
\arguments{
\item{seqtab}{DADA2 object. Sequence table, generated by `DADA2`.}

\item{taxa}{DADA2 object. Taxonomy table, generated by `DADA2`.}

\item{out_path}{Character. The absolute path of the output folder for saving DADA2 outputs.}
}
\value{
Three files are generated in the specified folder: `ASV_table.txt`, `ASV_taxa_table.txt`, and `ASV.fasta`.
}
\description{
A function to perform serial numbering for the sequence table generated by the DADA2 pipeline and save it in a specified folder for use in the PreLect process.
}
\examples{
# .....
seqtab <- makeSequenceTable(mergers)
seqtab.nochim <- removeBimeraDenovo(seqtab, method="consensus", multithread=TRUE, verbose=TRUE)

dim(seqtab.nochim)
sum(seqtab.nochim)/sum(seqtab)

 # Generate tracking information for filtering steps
getN <- function(x) sum(getUniques(x))
track <- cbind(out, sapply(dadaFs, getN), sapply(dadaRs, getN), sapply(mergers, getN), rowSums(seqtab.nochim))
#If processing a single sample, remove the sapply calls: e.g. replace sapply(dadaFs, getN) with getN(dadaFs)
colnames(track) <- c("input", "filtered", "denoisedF", "denoisedR", "merged", "nonchim")
rownames(track) <- sample.names
head(track)

# Assign taxonomy
taxa <- assignTaxonomy(seqtab.nochim, "~/tax/silva_nr_v132_train_set.fa.gz", multithread=TRUE)
taxa <- addSpecies(taxa, "~/tax/silva_species_assignment_v132.fa.gz")

output_path <- '~/dada2_output'
DADA2Adapter(seqtab.nochim, taxa, output_path)
dir(output_path)

data <- read.csv(paste0(output_path,'/ASV_table.txt'), sep = '\t')
taxa <- read.csv(paste0(output_path,'/ASV_taxa_table.txt'), sep = '\t')

}
