% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{evaluation}
\alias{evaluation}
\title{Feature Set Evaluation}
\usage{
evaluation(x_train, y_train, x_test, y_test, best_w, task)
}
\arguments{
\item{x_train}{Matrix. Training data with samples as rows and features as columns.}

\item{y_train}{Numeric vector. Labels for the training data.}

\item{x_test}{Matrix. Test data with samples as rows and features as columns.}

\item{y_test}{Numeric vector. Labels for the test data.}

\item{best_w}{Numeric vector. The feature weights estimated from PreLect, used to select relevant features.}

\item{task}{String. Specifies the type of task: either `classification` or `regression`.}
}
\value{
A list containing prediction performance metrics. For classification, it includes `AUC`; for regression, it includes `R2`.
}
\description{
Feature Set Evaluation
}
