% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.R
\name{generatePermutedTable}
\alias{generatePermutedTable}
\title{Sample randomization by fold change from label shuffling}
\usage{
generatePermutedTable(X, Y, control_sample)
}
\arguments{
\item{X}{Matrix or DataFrame. Raw count data with samples as rows and features as columns.}

\item{Y}{Character vector. Labels corresponding to the samples in the data.}

\item{control_sample}{Character. The name of the control sample.}
}
\value{
A DataFrame where the first column contains the observed order, and the following 999 columns contain randomized orders.
}
\description{
Sample randomization by fold change from label shuffling
}
