% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.R
\name{generatePermutedTableContinuous}
\alias{generatePermutedTableContinuous}
\title{Sample randomization by continuous label shuffling}
\usage{
generatePermutedTableContinuous(X, Y)
}
\arguments{
\item{X}{Matrix or DataFrame. Raw count data with samples as rows and features as columns.}

\item{Y}{Numeric vector. Labels corresponding to the samples in the data.}
}
\value{
A DataFrame where the first column contains the observed order, and the following 999 columns contain randomized orders.
}
\description{
Sample randomization by continuous label shuffling
}
