% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.R
\name{GetKOInfo}
\alias{GetKOInfo}
\title{Get the detail information of each KO from KEGG API (KEGGREST)}
\usage{
GetKOInfo(query_KOs)
}
\arguments{
\item{query_KOs}{Character vector. A list of quering KOs}
}
\value{
A DataFrame with gene name, symbol, involved pathway for each KO.
}
\description{
Get the detail information of each KO from KEGG API (KEGGREST)
}
\examples{

Descrip <- read.csv('.../KO/pred_metagenome_unstrat_descrip.tsv', as.is = TRUE, stringsAsFactors = F, sep = '\t', header = T)
query_KOs <- unique(Descrip$function.)
KOinfo <- GetKOInfo(query_KOs)

}
