% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.R
\name{GSEATestwithFC}
\alias{GSEATestwithFC}
\title{GSEA test with binary labels}
\usage{
GSEATestwithFC(KO_index, X, Y, case_sample)
}
\arguments{
\item{KO_index}{DataFrame. A table containing pair-relationship between KOs and taxa, generally is `pred_metagenome_contrib.tsv` file generated` by `PICRUst2`}

\item{X}{Matrix or DataFrame. Raw count data with samples as rows and features as columns.}

\item{Y}{Character vector. Labels corresponding to the samples in the data.}

\item{case_sample}{Character. The name of the case sample.}
}
\value{
A list containing two DataFrames:
- `Actived_KO`: Results of the GSEA analysis for case-activated KOs, including enrichment scores (ES), z-values, and p-values for each KO.
- `Suppressed_KO`: Results of the GSEA analysis for case-suppressed KOs, including enrichment scores (ES), z-values, and p-values for each KO.
}
\description{
GSEA test with binary labels
}
\examples{

KOindex <- read.table(".../KO/pred_metagenome_contrib.tsv", sep = "\t", header = TRUE)
GSEAresult <- GSEATestwithFC(KOindex, ASV_table, groupings$Class, "Cancer")
Actived_result <- GSEAresult$Actived_KO
Actived_result <- Actived_result[Actived_result$z > 2,]
nrow(Actived_result)

}
