% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.R
\name{PathwayEnrichment}
\alias{PathwayEnrichment}
\title{Fisher's exact test for pathway enrichment}
\usage{
PathwayEnrichment(filtered_KO, KOmap)
}
\arguments{
\item{filtered_KO}{Character vector. A list of significant KOs obtained from the GSEA test.}

\item{KOmap}{KO-pathway mapping information generated by the `GetKOInfo` function using the KEGG API.}
}
\value{
DataFrame. A data frame containing the pathway enrichment results.
}
\description{
Fisher's exact test for pathway enrichment
}
