% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.R
\name{permutatedTest}
\alias{permutatedTest}
\title{Permutation test function using the GSEA method}
\usage{
permutatedTest(permutated_table, ko_mapper)
}
\arguments{
\item{permutated_table}{DataFrame. A table containing randomized order, generated by `generatePermutedTable` or `generatePermutedTableContinuous`.}

\item{ko_mapper}{List. A list that records taxa groups for each KO, generated by `map_ko_to_taxon`.}
}
\value{
A DataFrame containing the results of the permutation test, including enrichment score (ES), z-value, and p-value for each KO.
}
\description{
Permutation test function using the GSEA method
}
