% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiClass.R
\name{PreLectMultiClass}
\alias{PreLectMultiClass}
\title{Multi-Class Feature Selection with PreLect}
\usage{
PreLectMultiClass(
  X_scale,
  X_raw,
  Y,
  lambda,
  run_echo = FALSE,
  max_iter = 10000,
  tol = 1e-04,
  lr = 0.001,
  alpha = 0.9,
  epsilon = 1e-08
)
}
\arguments{
\item{X_scale}{Matrix or DataFrame. Scaled data with samples as rows and features as columns, used for machine learning. If no scaled data is provided, raw count data may be used.}

\item{X_raw}{Matrix or DataFrame. Raw count data with samples as rows and features as columns, used to calculate feature prevalence.}

\item{lambda}{Numeric. Lambda value, the intensity of regularization. Strongly suggested to be determined by the `LambdaTuningMultiClass` process.}

\item{run_echo}{Logical. If TRUE, prints the training result (default is FALSE).}

\item{max_iter}{Integer. Maximum number of iterations taken for the solvers to converge (default is 10000).}

\item{tol}{Numeric. Tolerance for stopping criteria (default is 1e-4).}

\item{lr}{Numeric. Learning rate in RMSprop optimizer (default is 0.001).}

\item{alpha}{Numeric. Smoothing constant in RMSprop optimizer (default is 0.9).}

\item{epsilon}{Numeric. Small constant added to the denominator to improve numerical stability (default is 1e-8).}
}
\value{
A list containing:
- `coef_table`: A data frame with class-specific coefficients for each feature, including both coefficient values and class tendency.
- `loss_value`: Final loss value achieved during training.
- `convergence`: Difference between the final and penultimate iterations, indicating convergence level.
}
\description{
`PreLectMultiClass` performs feature selection for multi-class classification by identifying class-specific coefficients based on input lambda regularization.
}
\examples{
set.seed(42)
n_samples <- 10
n_features <- 100

X_raw <- matrix(rnbinom(n_features * n_samples, size = 10, mu = 1), nrow = n_features, ncol = n_samples)
X_scaled <- t(scale(t(X_raw)))  # feature-wise z-standardization
diagnosis <- c('CRC','CRC','control','Adenoma','Adenoma','CRC','control','control','CRC','CRC')

result <- PreLectMultiClass(X_scaled, X_raw, diagnosis, lambda = 0.01)

}
