% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreLectR-package.R
\docType{package}
\name{PreLectR-package}
\alias{PreLectR}
\alias{PreLectR-package}
\title{PreLectR: Feature Engineering in Sparse Matrices}
\description{
Implement the PreLect algorithm, integrating L1 regularization with an inverted prevalence penalty, to select universal and informative features in sparse data, particularly for microbiome studies.
}
\keyword{internal}
