% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{TaxaProperty}
\alias{TaxaProperty}
\title{Taxonomy tendency visualization}
\usage{
TaxaProperty(feat_prop, taxa_table, taxa_level = "Order", pvl_filter = 0.4)
}
\arguments{
\item{feat_prop}{DataFrame. A data frame containing feature properties, generated by `FeatureProperty`.}

\item{taxa_table}{DataFrame. A data frame containing taxonomy information for each feature, generated by `DADA2Adapter`.}

\item{taxa_level}{Character. Specifies the column name in `taxa_table` that corresponds to the desired taxonomy level for analysis (default is "Order").}

\item{pvl_filter}{Numeric. A threshold for the prevalence of taxonomy units. Taxa with prevalence below this threshold will be excluded (default is 0.4).}
}
\value{
A list containing two elements:
- `effectSizePlot`: A ggplot object visualizing the effect size of each taxonomy unit.
- `selectedInfo`:A data frame summarizing the information for each selected feature, including coefficients and taxonomy assignments.
}
\description{
This function visualizes the properties of taxonomy derived from selected features and generates a plot summarizing the effect size of each taxonomy unit.
}
\examples{

set.seed(42)
n_samples <- 10
n_features <- 100

X_raw <- matrix(rnbinom(n_features * n_samples, size = 10, mu = 1), nrow = n_features, ncol = n_samples)
rownames(X_raw) <- paste0('feat',1:n_features)
X_scaled <- t(scale(t(X_raw)))  # feature-wise z-standardization
diagnosis <- c('CRC','CRC','health','CRC','health','CRC','health','health','CRC','CRC')
diagnosis <- factor(diagnosis, levels=c('health', 'CRC')) # assign 'health' as control sample
pvlvec <- GetPrevalence(X_raw)

result <- PreLect(X_scaled, pvlvec, diagnosis, lambda=1e-4, task="classification")
prop <- FeatureProperty(X_raw, diagnosis, result)

TaxaProperty(prop, taxa, taxa_level = "Order", pvl_filter = 0.4)

}
