% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{TrainTextSplit}
\alias{TrainTextSplit}
\title{Data Splitting}
\usage{
TrainTextSplit(labels, ratio = 0.7, task = "classification")
}
\arguments{
\item{labels}{Vector. Labels for the data.}

\item{ratio}{Numeric. The proportion of samples to assign to the training set (default is 0.7).}

\item{task}{String. Specifies the type of task: either `classification` or `regression` (default is `classification`).}
}
\value{
A list with two elements: `train_idx` for training indices and `test_idx` for testing indices.
}
\description{
Splits data into stratified training and testing sets based on specified proportions for each class.
}
