% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{GetPrevalence}
\alias{GetPrevalence}
\title{Get Prevalence}
\usage{
GetPrevalence(X)
}
\arguments{
\item{X}{Matrix or DataFrame. Raw count data with samples as rows and features as columns.}
}
\value{
A vector of prevalence for each feature.
}
\description{
Calculate prevalence of each feature.
}
